Card List Added
Closes #7
Added a card list to the code, and added public/cardLists directory for
future lists as well. For now, the list is just card ids, each on its
own line. Multiple of a card would mean multiple entries. Will have to
do work in the future to allow the cubes in csv form to be imported.
Also added a shuffle function, and the server will shuffle the deck
right before setting the data card list.