Setting Up Manual End State
When all players have at least 40 cards, the players should be able to
quit whenever they please. Added an end game button to the
liquidate/wager screens, and methods for toggling whether it is being
displayed or not. These methods are not currently being implemented.

Added a status to players on whether they have requested the game to end
or not, and methods for checking if all have requested it. Updated what
happens when switching game states: before only the "impacted" status
was reset (liquidated reset when liquidating was over, wagered reset
when wagering was over), now it resets all of the player statuses, both
to catch any weird edge cases where something was true when it shouldn't
be, as well as to additionally reset requestedEnd.

Added a screen to the client that just has a download button, and made
the method(s) for the server to tell the clients to go there.