Server Function to Update End Buttons
When the number of cards player(s) have can change (wagering completed
and liquidating), whether or not the end buttons are being shown can
change. The base criteria to show the button is that every player has at
least 40 cards. Added a function that will tell clients to show the buttons if
before the change it wasn't, and now it should be, and to hide the
buttons if before the change it was, and now it shouldn't be. That way
it won't constantly be emitting to hide buttons while players climb to
40 cards, and constantly emitting to show buttons when players all have
40 and keep playing.