Allow Current Card to Change
Closes #13
Now the current card is set when the game is started, and every time a
wager round is completed, a new one is set by popping data.cardList.
Game will also check if there are no more cards remaining before
performing this pop, and if there aren't it will send all players to the
end screen.