Card Sizing and Section Labels
Cards within the deck now have a deck-card class, and a defined width so
that a certain number fit in a row, as well as making it so when it is
selected and border added, they remain the same size. Also added
"Wagered Card" and "Hovered Card" headers to the relevant sections.