Wager/End Button Bug Fix
After testing with Christian, there was an issue in the server with
players' wagered status not being reset, and end buttons not appearing.
The wager issue was just that resetLiquidated being used twice instead
of resetLiquidated and resetWagered. The end button issue was even
simpler, attempting to change the style of endButtons instead of
endButton.