Limited All Liquidated Emits
Closes #27
Added a liquidatedAtLeastOneCard property to player, that is true if a
player clicks the liquidate button and had selected at least one card,
and the server was able to liquidate those card(s). Used this when
running the all liquidated actions to see if a player actually needed a
token/deck/deck size update.