Add Player Statuses Object
Closes #28
Added a statuses object to player, so that adding any statuses in the
future will be able to just add without having to update
resetAllStatuses. Also made a function for setting this object to all
properties set to false, both to initialize the object, as well as to
make it easy to reset.