Showing Player's Wager on Hover
Closes #11
Added some lines to the generateCardElement so that the wager can be
extracted from the cardObject, and so that when a card is hovered it
will set the newly created previous-wager <p> elements to what the
player wagered for that card, and make it blank when the mouse leaves
that card. Server allLiquidated code was updated so that currentWagers
is reset after the player is given the card they just won.