Added Ability to use Enter Key
On the wagering screen, the text box will now automatically be the focus
every time it shows. Clicking enter while in that box will submit the
wager. On the liquidating screen, the liquidate button is focused when
it shows, clicking card(s) will take that focus away, but at least if
you want to not liquidate you can quickly click enter.