Allow Wagered Description Scroll
Closes #32
The wagered cards description now is able to be scrolled if it overflows
the left column. Also lowered the size of the description text so most
fit anyways. Hovered card section now also allows the overflow, but you
can't really scroll the whole section because you are hovering over a
card.